requirements.txt
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-WTF==1.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
WTForms==3.0.1
Werkzeug==2.1.2
click==8.1.3
dominate==2.6.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
pip==22.0.4
python-dotenv==0.20.0
setuptools==62.1.0
visitor==0.1.3
zipp==3.8.0